requirements.txt
Pillow==4.3.0
scipy==1.0.0
torch==0.3.0.post4
torchvision==0.2.0
